# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Test app | Secunda test",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/organizations/activities": {
            "get": {
                "tags": [
                    "Organization Activities"
                ],
                "summary": "Получить список активностей организаций",
                "operationId": "c70a1186e9937a43c86c90a49c5fc04f",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 20,
                            "maximum": 1000,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/OrganizationActivity"
                                            }
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PaginationMeta"
                                        },
                                        "links": {
                                            "$ref": "#/components/schemas/PaginationLinks"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Organization Activities"
                ],
                "summary": "Создать новую активность организации",
                "operationId": "430d14d3198f31c1c0936adef6e0cafb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OrganizationActivityCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Активность организации создана",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/OrganizationActivity"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/organizations/activities/{organizationActivity}": {
            "get": {
                "tags": [
                    "Organization Activities"
                ],
                "summary": "Получить активность организации по ID",
                "operationId": "f2e57254951cb30b4c00ec5a257a8b4c",
                "parameters": [
                    {
                        "name": "organizationActivity",
                        "in": "path",
                        "description": "ID активности организации",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/OrganizationActivity"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Активность организации не найдена"
                    }
                }
            },
            "put": {
                "tags": [
                    "Organization Activities"
                ],
                "summary": "Обновить активность организации",
                "operationId": "b76a3aafde4913c9aeb3ae069125d684",
                "parameters": [
                    {
                        "name": "organizationActivity",
                        "in": "path",
                        "description": "ID активности организации",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OrganizationActivityUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Активность организации обновлена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/OrganizationActivity"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Активность организации не найдена"
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Organization Activities"
                ],
                "summary": "Удалить активность организации",
                "operationId": "0d55f86763b54e93bf315d29408c6228",
                "parameters": [
                    {
                        "name": "organizationActivity",
                        "in": "path",
                        "description": "ID активности организации",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Активность организации удалена"
                    },
                    "404": {
                        "description": "Активность организации не найдена"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "OrganizationActivity": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "organization_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "activity_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "OrganizationActivityCreate": {
                "required": [
                    "organization_id",
                    "activity_id"
                ],
                "properties": {
                    "organization_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "activity_id": {
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
            "OrganizationActivityUpdate": {
                "properties": {
                    "organization_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "activity_id": {
                        "type": "integer",
                        "example": 2
                    }
                },
                "type": "object"
            },
            "PaginationMeta": {
                "properties": {
                    "current_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "from": {
                        "type": "integer",
                        "example": 1
                    },
                    "last_page": {
                        "type": "integer",
                        "example": 10
                    },
                    "path": {
                        "type": "string",
                        "example": "http://localhost/api/buildings"
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 20
                    },
                    "to": {
                        "type": "integer",
                        "example": 20
                    },
                    "total": {
                        "type": "integer",
                        "example": 200
                    }
                },
                "type": "object"
            },
            "PaginationLinks": {
                "properties": {
                    "first": {
                        "type": "string",
                        "example": "http://localhost/api/buildings?page=1"
                    },
                    "last": {
                        "type": "string",
                        "example": "http://localhost/api/buildings?page=10"
                    },
                    "prev": {
                        "type": "string",
                        "example": "http://localhost/api/buildings?page=1"
                    },
                    "next": {
                        "type": "string",
                        "example": "http://localhost/api/buildings?page=2"
                    }
                },
                "type": "object"
            },
            "ValidationError": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "data": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "apiKeyAuth": {
                "type": "apiKey",
                "description": "Постоянный API ключ для доступа к API",
                "name": "X-API-Key",
                "in": "header"
            }
        }
    },
...
}